requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
pymupdf
python-docx
